Move patch out to dns module
